Pharmacy_Inventory_API/requirements.txt
django
djangorestframework
django-crispy-forms
django-filter
django-cors-headers
djangorestframework-simplejwt 
mysql
mysqlclient
python-dotenv
Pillow
drf-yasg
psycopg2-binary 
gunicorn
whitenoise
pytest
pytest-django
pytest-cov

